# docs/swagger.yaml
basePath: /api/v1
definitions:
  gopet_internal_core_transport_response.ErorrResponse:
    properties:
      error:
        example: full error text
        type: string
      message:
        example: short human-readable message
        type: string
    type: object
...
  internal_features_users_transort_http.CreateUserRequest:
    properties:
      full_name:
        example: Ivan Ivanon
        maxLength: 100
        minLength: 3
        type: string
      phone_number:
        example: "+38097123804"
        maxLength: 15
        minLength: 10
        type: string
    required:
    - full_name
    type: object
  internal_features_users_transort_http.CreateUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_users_transort_http.GetUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_users_transort_http.PatchUserRequest:
    properties:
      full_name:
        example: Iv1n Invaov
        type: string
      phone_number:
        example: "+38097223804"
        type: string
    type: object
  internal_features_users_transort_http.PatchUserResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
  internal_features_users_transort_http.UserDTOResponse:
    properties:
      full_name:
        example: Ivan Ivanov
        type: string
      id:
        example: 10
        type: integer
      phone_number:
        example: "+38097123804"
        type: string
      version:
        example: 2
        type: integer
    type: object
host: 172.25.87.4:5050
info:
  contact: {}
  description: gopet rest api schema
  title: Golang GOPET API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Get existing users in system with optional pagination
      parameters:
      - description: Pagination limit parameter
        in: query
        name: limit
        type: integer
      - description: Pagination offset parameter
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users
          schema:
            items:
              $ref: '#/definitions/internal_features_users_transort_http.UserDTOResponse'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Get Users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create new user in system
      parameters:
      - description: CreateUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transort_http.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            $ref: '#/definitions/internal_features_users_transort_http.CreateUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Create User
      tags:
      - users
  /users/{id}:
    delete:
      description: Delete existing user in system
      parameters:
      - description: ID of the user to be deleted
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted user
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Delete User
      tags:
      - users
    get:
      description: Get existing user in system
      parameters:
      - description: ID of the user to be retrieved
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user
          schema:
            $ref: '#/definitions/internal_features_users_transort_http.GetUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Get User
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: |-
        ### Three-state logic:
        1. **Field is not provided:** nothing to be done
        2. **Field is provided(+12390481234):** creating new
        3. **Field is provided(null):** set to null
      parameters:
      - description: ID of the user to be patched
        in: path
        name: id
        required: true
        type: integer
      - description: PatchUser request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_features_users_transort_http.PatchUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully patched user
          schema:
            $ref: '#/definitions/internal_features_users_transort_http.PatchUserResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/gopet_internal_core_transport_response.ErorrResponse'
      summary: Patch User
      tags:
      - users
swagger: "2.0"